+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Latest
+
+- [Bugfix] Fix missing webui env directory
+
 ## 3.1.0 (2019-03-09)
 
 - [Improvement] Install python requirements in virtual env in docker image